requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-multipart
pydantic
aiofiles